src/ssmai_backend/schemas/products_schemas.py
from datetime import datetime

from pydantic import BaseModel, ConfigDict


class ProductSchema(BaseModel):
    titulo: str
    preco: float
    quantidade: int
    categoria: str
    status: str  # TODO: Migrar para Enum


class PublicProductSchema(ProductSchema):
    id: int
    created_at: datetime
    updated_at: datetime
    model_config = ConfigDict(from_attributes=True)


class ProductsList(BaseModel):
    products: list[PublicProductSchema]
