pyproject.toml
[project]
name = "ssmai-backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Joao_SDAVieira"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard] (>=0.116.1,<0.117.0)",
    "psycopg[binary] (>=3.2.9,<4.0.0)",
    "sqlalchemy[asyncio] (>=2.0.43,<3.0.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
    "fastapi-users[sqlalchemy] (>=14.0.1,<15.0.0)",
]

[tool.poetry.dependencies]
python = "^3.13"


[tool.poetry]
packages = [{include = "ssmai_backend", from = "src"}]


[tool.ruff]
line-length = 79
extend-exclude = ['migrations']


[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT', 'FAST']


[tool.ruff.format]
preview = true


[tool.taskipy.tasks]
run = 'fastapi dev src/ssmai_backend/app.py'
test =  'pytest --cov=src -vv'
post_test = 'coverage html'
lint = 'ruff check . && ruff check . --diff'
format = 'ruff check . --fix && ruff format .'
pre_test = 'task lint'


[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'
asyncio_default_fixture_loop_scope = 'function'


[tool.poetry.group.dev.dependencies]
ruff = "^0.12.12"
pytest = "^8.4.2"
pytest-cov = "^6.2.1"
taskipy = "^1.14.1"
pytest-asyncio = "^1.1.0"
testcontainers = "^4.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
concurrency = ["thread", "greenlet"]